LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/location-href-intercept_currententrychange-expected.txt


FAIL event and promise ordering for location.href setter intercepted by intercept() assert_array_equals: expected property 0 to be "navigate" but got "promise microtask" (expected array ["navigate", "currententrychange", "handler run", "promise microtask", "navigatesuccess", "transition.finished fulfilled"] got ["promise microtask", "navigate", "currententrychange", "handler run", "navigatesuccess", "transition.finished fulfilled"])
